Crate icu_provider_uprops

icu_provider_uprops contains implementations of the ICU4X DataProvider interface backed by TOML files exported by the ICU4C icuwriteuprops tool. Create a directory containing TOML files for the necessary Unicode properties and then pass the path into the PropertiesDataProvider.

Important: This data provider implementation is not optimized for production use. It is much more efficient if you use FsDataProvider or StaticDataProvider instead.

Structs

EnumeratedPropertyCodePointTrieProvider

This data provider returns CodePointTrie data inside a UnicodePropertyMap data struct. The source data is the same as that of crate::provider::PropertiesDataProvider, which is a TOML file of data for the property(-ies) desired, as given by the ICU4C property data exporter tool.

PropertiesDataProvider

This data provider returns UnicodeSet data inside a UnicodeProperty data struct. The source data is in the form of a directory of TOML file(s) of data for the property(-ies) desired, as given by the ICU4C property data exporter tool.